+++ src/gui/src/components/explorer-grid/results/empty-results-state.tsx
