+++ src/gui/src/components/explorer-grid/results/index.tsx
@@ -1,10 +1,10 @@
-import { ResultItem } from '@/components/explorer-grid/result-item.tsx'
+import { ResultItem } from '@/components/explorer-grid/results/result-item.tsx'
 import type { GridItemData } from '@/components/explorer-grid/types.ts'
 import { GridHeader } from '@/components/explorer-grid/header.tsx'
-import { EmptyResultsState } from '@/components/explorer-grid/empty-results-state.tsx'
+import { EmptyResultsState } from '@/components/explorer-grid/results/empty-results-state.tsx'
 import { Badge } from '@/components/ui/badge.tsx'
 
-export const GridResults = ({ items }: { items: GridItemData[] }) => {
+export const Results = ({ items }: { items: GridItemData[] }) => {
   if (!items.length) {
     return <EmptyResultsState />
   }
